Introduce latest_valuation
This commit adds logic to figure out what the latest valuation for a neighborhood is.

The approach that we went with is to try to find a date when all the
properties a neighborhood has have a valuation. Some things to look out
for include:

Hood 1:
  - House 1 (created 04-01), last valuated (05-01)
  - House 2 (created 04-01), last valuated (05-01)
  - House 3 (created 04-01), last valuated (05-01)
  - House 4 (created 05-01), not valuated

If we invoke `latest_valuation` after `05-01`, we will not receive an
amount because house 4 has not yet been valuated.

This is a rigorous approach, but it's the simplest for now.