Bug: always compare date objects
There was a bug in the valuation_before method where if a datetime
object was passed in as the start_date, you could accidentally get that
date.

E.g. valuation_before(<Today>) # => could return a result for today

This was causing us to adjust our date ranges by adding an extra day.