Convert Daily Diff calculator to generic diff calculator (#42)
* Introduce date-agnostic difference calculator

Rather than relying on a time-specific calculator (e.g. `daily`,
potentially `monthly`, etc), we use a single calculator that will
calculate differences regardless of date ranges.

* Test out the difference calculator

All the business logic for generating differences is now largely being
moved into the DifferenceCalculator. Therefore, we should test out the
class itself, so that we don't end up with massive unit tests for other
things that rely on it.

* Add specs for the difference calculator

* Simplify Slack Client to use difference calculator

* Use a renderer to contain logic for rendering a diff

* Add a way to send monthly price summary

* Remove daily diff calculator

* Include NumberHelper and DateHelper instead of extending them so their helper methods can be used directly in the class

* Bug: always compare date objects

There was a bug in the valuation_before method where if a datetime
object was passed in as the start_date, you could accidentally get that
date.

E.g. valuation_before(<Today>) # => could return a result for today

This was causing us to adjust our date ranges by adding an extra day.

* Introduce a `valuation_on_or_before` method

This commit adds a new method that allows us to more easily express an
inclusive date range.

* Remove defaults in hood difference

Co-authored-by: Hebron George <hebrontgeorge@gmail.com>