Ensure monthly task only runs once a month
This commit adds a Rails cache method to check if a message was produced
during the run.

This approach was taken due to how low impact/how quickly it could be
deployed.

Another approach is to use a database model and only publish it if it
doesn't exist (which would help do some monthly analytics if we wanted).